+++ packages/example/src/commands/mod.rs
@@ -1,0 +1,3 @@
+pub mod add_element_command;
+pub mod remove_element_command;
+pub mod update_element_command;