+++ packages/example/src/services/mutex/ports/mod.rs
@@ -1,2 +1,0 @@
-pub mod mutex_repository_port;
-pub mod mutex_service_port;