+++ packages/example/src/services/user_service/mod.rs
@@ -1,4 +1,0 @@
-pub mod create_user_command;
-pub mod get_user_by_name_query;
-pub mod update_user_command;
-pub mod user_service;